+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to
 [Compatible Versioning](https://github.com/staltz/comver).
 
 ## [Unreleased]
+
+## [0.4] - 2019-01-09
 ### Added
 - Schema `enum` declaration.
 - `docIdFromUuid`.
@@ -62,7 +64,8 @@ and this project adheres to
   - RON-Schema
   - RON-Schema TemplateHaskell code generator
 
-[Unreleased]: https://github.com/ff-notes/ff/compare/v0.3...HEAD
+[Unreleased]: https://github.com/ff-notes/ff/compare/v0.4...HEAD
+[0.4]: https://github.com/ff-notes/ff/compare/v0.3...v0.4
 [0.3]: https://github.com/ff-notes/ff/compare/v0.2...v0.3
 [0.2]: https://github.com/ff-notes/ff/compare/v0.1...v0.2
 [0.1]: https://github.com/ff-notes/ff/tree/v0.1